+++ frontend/e-commerce-service/src/app/components/review-management/review-creation/review-creation.component.ts
@@ -1,0 +1,12 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-review-creation',
+  standalone: true,
+  imports: [],
+  templateUrl: './review-creation.component.html',
+  styleUrl: './review-creation.component.css'
+})
+export class ReviewCreationComponent {
+
+}